requirements.txt
python-telegram-bot
pytest
pytest-mock
pytest-cov
pylint
telethon